refactored frontend naming
