--- frontend/src/app/screens/landing-screen/landing-screen.component.spec.ts
@@ -1,25 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { LandingScreenComponent } from './landing-screen.component';
-
-describe('LandingScreenComponent', () => {
-  let component: LandingScreenComponent;
-  let fixture: ComponentFixture<LandingScreenComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      declarations: [ LandingScreenComponent ]
-    })
-    .compileComponents();
-  });
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(LandingScreenComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
