--- frontend/src/app/screens/price-is-nice-screen/components/babo-result/babo-result.component.spec.ts
@@ -1,25 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { BaboResultComponent } from './babo-result.component';
-
-describe('BaboResultComponent', () => {
-  let component: BaboResultComponent;
-  let fixture: ComponentFixture<BaboResultComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      declarations: [ BaboResultComponent ]
-    })
-    .compileComponents();
-  });
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(BaboResultComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
