--- frontend/src/app/screens/price-is-nice-screen/components/guess-information/guess-information.component.spec.ts
@@ -1,25 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { GuessInformationComponent } from './guess-information.component';
-
-describe('GuessInformationComponent', () => {
-  let component: GuessInformationComponent;
-  let fixture: ComponentFixture<GuessInformationComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      declarations: [ GuessInformationComponent ]
-    })
-    .compileComponents();
-  });
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(GuessInformationComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
