--- frontend/src/app/screens/price-is-nice-screen/components/guess-input/guess-input.component.spec.ts
@@ -1,25 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { GuessInputComponent } from './guess-input.component';
-
-describe('GuessInputComponent', () => {
-  let component: GuessInputComponent;
-  let fixture: ComponentFixture<GuessInputComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      declarations: [ GuessInputComponent ]
-    })
-    .compileComponents();
-  });
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(GuessInputComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
